New hwaddr type
This type stores the hwaddr as an int internally, so it avoids case problems
in strings and when comparing between two hwaddr(already converted) will be
faster than comparing strings.

However when comparing with a string, it will be slower because it will need
to convert the string representation to int to compare.