BUG: When a device was not pinged, a skew would happen.
Basically the code was presuming a 1 to 1 match between pings and
devices, however when some devices are not found in the ARP table
they will not be pinged, so the assumption will not hold any more
and the code would apply results of the ping to the wrong device.

This commit introduces a queue to keep track of the match in any
case instead.